Add upload files about DQN network
The files is our RL DQN network, I found that a lot of financial stock trading uses the DQN algorithm.  I used a double DQN based on some tutorials and the basic DQN algorithm, which achieves just the right amount of convergence.I think this algorithm should be able to achieve some results and of course can be our alternative algorithm if you have a better.